// src/interfaces/AggregateAdp.ts
export interface AggregateAdp {
    sleeperStandard: number;
    espnStandard: number;
    fantraxStandard: number;
    nffcStandard: number;
    underdogStandard: number;
    sleeperHalfPpr: number;
    espnHalfPpr: number;
    fantraxHalfPpr: number;
    nffcHalfPpr: number;
    underdogHalfPpr: number;
    sleeperPpr: number;
    espnPpr: number;
    fantraxPpr: number;
    nffcHpr: number;
    underdogPpr: number;
    sleeper2Qb: number;
    espnH2Qb: number;
    fantrax2Qb: number;
    nffc2Qb: number;
    underdog2Qb: number;
    sleeperDynastyStandard: number;
    espnDynastyStandard: number;
    fantraxDynastyStandard: number;
    nffcDynastyStandard: number;
    underdogDynastyStandard: number;
    sleeperDynastyHalfPpr: number;
    espnDynastyHalfPpr: number;
    fantraxDynastyHalfPpr: number;
    nffcDynastyHalfPpr: number;
    underdogDynastyHalfPpr: number;
    sleeperDynastyPpr: number;
    espnDynastyPpr: number;
    fantraxDynastyPpr: number;
    nffcDynastyHpr: number;
    underdogDynastyPpr: number;
    sleeperDynasty2Qb: number;
    espnDynasty2Qb: number;
    fantraxDynasty2Qb: number;
    nffcDynasty2Qb: number;
    underdogDynasty2Qb: number;
}